+++ packages/button/src/styles.ts
@@ -330,128 +330,3 @@ export function getClassName({
     size,
   );
 }
-
-const iconColor: Record<Mode, Record<Variant, string>> = {
-  [Mode.Light]: {
-    [Variant.Primary]: css`
-      color: ${uiColors.green.light3};
-    `,
-    [Variant.PrimaryOutline]: css`
-      color: ${uiColors.green.dark1};
-    `,
-    [Variant.Default]: css`
-      color: ${uiColors.gray.dark1};
-    `,
-    [Variant.Danger]: css`
-      color: ${uiColors.red.light3};
-    `,
-    [Variant.DangerOutline]: css`
-      color: ${uiColors.red.base};
-    `,
-  },
-
-  [Mode.Dark]: {
-    [Variant.Primary]: css`
-      color: ${uiColors.green.light2};
-    `,
-    [Variant.PrimaryOutline]: css`
-      color: #13aa52;
-    `,
-    [Variant.Default]: css`
-      color: ${uiColors.gray.light1};
-    `,
-    [Variant.Danger]: css`
-      color: ${uiColors.red.light3};
-    `,
-    [Variant.DangerOutline]: css`
-      color: #f97216;
-    `,
-  },
-};
-
-const onlyIconColor: Record<Mode, Record<Variant, string>> = {
-  [Mode.Light]: {
-    [Variant.Primary]: css`
-      color: ${uiColors.white};
-    `,
-    [Variant.PrimaryOutline]: css`
-      color: ${uiColors.green.dark1};
-    `,
-    [Variant.Default]: css`
-      color: ${uiColors.gray.dark2};
-    `,
-    [Variant.Danger]: css`
-      color: ${uiColors.white};
-    `,
-    [Variant.DangerOutline]: css`
-      color: #cf4a22;
-    `,
-  },
-  [Mode.Dark]: {
-    [Variant.Primary]: css`
-      color: ${uiColors.white};
-    `,
-    [Variant.PrimaryOutline]: css`
-      color: #0ad05b;
-    `,
-    [Variant.Default]: css`
-      color: ${uiColors.white};
-    `,
-    [Variant.Danger]: css`
-      color: ${uiColors.white};
-    `,
-    [Variant.DangerOutline]: css`
-      color: #f97216;
-    `,
-  },
-};
-
-const iconSize: Record<Size, string> = {
-  [Size.XSmall]: css`
-    height: 14px;
-    width: 14px;
-  `,
-  [Size.Small]: css`
-    height: 16px;
-    width: 16px;
-  `,
-  [Size.Default]: css`
-    height: 16px;
-    width: 16px;
-  `,
-  [Size.Large]: css`
-    height: 20px;
-    width: 20px;
-  `,
-};
-
-const disabledIconStyle: Record<Mode, string> = {
-  [Mode.Light]: css`
-    color: ${uiColors.gray.light1};
-  `,
-  [Mode.Dark]: css`
-    color: ${uiColors.gray.dark1};
-  `,
-};
-
-type IconStyle = Required<
-  Pick<ButtonProps, 'variant' | 'size' | 'darkMode' | 'disabled'>
-> & {
-  isIconOnlyButton: boolean;
-};
-
-export function getIconStyle({
-  variant,
-  size: sizeProp,
-  darkMode,
-  disabled,
-  isIconOnlyButton,
-}: IconStyle) {
-  const mode = darkMode ? Mode.Dark : Mode.Light;
-  const color = isIconOnlyButton
-    ? onlyIconColor[mode][variant]
-    : iconColor[mode][variant];
-  const size = iconSize[sizeProp];
-
-  return cx(color, { [disabledIconStyle[mode]]: disabled }, size);
-}
